# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "AirGo前后分离,多用户代理面板",
        "title": "AirGo",
        "contact": {
            "url": "https://github.com/pura-panel/airgo-panel"
        },
        "license": {
            "name": "GPL v3.0",
            "url": "https://github.com/pura-panel/airgo-panel/blob/main/LICENSE"
        }
    },
    "paths": {
...
        "/api/customer/ticket/firstTicket": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[customer api] ticket"
                ],
                "summary": "获取工单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "description": "参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Ticket"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功；正常：业务代码 code=0；错误：业务代码code=1",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseStruct"
                        }
                    }
                }
            }
        },
        "/api/customer/ticket/getUserTicketList": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[customer api] ticket"
                ],
                "summary": "获取工单列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "description": "参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.QueryParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功；正常：业务代码 code=0；错误：业务代码code=1",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseStruct"
                        }
                    }
                }
            }
        },
        "/api/customer/ticket/newTicket": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[customer api] ticket"
                ],
                "summary": "新建工单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "description": "参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Ticket"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功；正常：业务代码 code=0；错误：业务代码code=1",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseStruct"
                        }
                    }
                }
            }
        },
        "/api/customer/ticket/sendTicketMessage": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[customer api] ticket"
                ],
                "summary": "发送工单消息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "description": "参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.TicketMessage"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功；正常：业务代码 code=0；错误：业务代码code=1",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseStruct"
                        }
                    }
                }
            }
        },
        "/api/customer/ticket/updateUserTicket": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[customer api] ticket"
                ],
                "summary": "更新工单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "description": "参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Ticket"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功；正常：业务代码 code=0；错误：业务代码code=1",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseStruct"
                        }
                    }
                }
            }
        },
        "/api/customer/traffic/getSubTrafficList": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[customer api] traffic"
                ],
                "summary": "获取订阅流量记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "description": "参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CustomerService"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功；正常：业务代码 code=0；错误：业务代码code=1",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseStruct"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.CustomerService": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "des": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                },
                "goods_id": {
                    "description": "OriginalAmount string `json:\"original_amount\"  gorm:\"comment:原始金额\"`\n商品参数-基础参数",
                    "type": "integer"
                },
                "goods_type": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_renew": {
                    "description": "续费参数",
                    "type": "boolean"
                },
                "node_connector": {
                    "type": "integer"
                },
                "node_speed_limit": {
                    "type": "integer"
                },
                "price": {
                    "type": "string"
                },
                "renewal_amount": {
                    "type": "string"
                },
                "service_end_at": {
                    "type": "string"
                },
                "service_start_at": {
                    "type": "string"
                },
                "service_status": {
                    "type": "boolean"
                },
                "sub_status": {
                    "type": "boolean"
                },
                "sub_uuid": {
                    "type": "string"
                },
                "subject": {
                    "type": "string"
                },
                "total_bandwidth": {
                    "description": "商品参数-订阅",
                    "type": "integer"
                },
                "traffic_reset_day": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "used_down": {
                    "type": "integer"
                },
                "used_up": {
                    "type": "integer"
                },
                "user_id": {
                    "description": "关联用户",
                    "type": "integer"
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
...
        "model.FieldParamsItem": {
            "type": "object",
            "properties": {
                "condition": {
                    "description": "= \u003e \u003c \u003c\u003e like",
                    "type": "string"
                },
                "condition_value": {
                    "type": "string"
                },
                "field": {
                    "type": "string"
                },
                "field_type": {
                    "type": "string"
                },
                "operator": {
                    "description": "AND OR",
                    "type": "string"
                }
            }
        },
...
        "model.Pagination": {
            "type": "object",
            "properties": {
                "order_by": {
                    "type": "string"
                },
                "page_num": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                }
            }
        },
...
        "model.QueryParams": {
            "type": "object",
            "required": [
                "table_name"
            ],
            "properties": {
                "field_params_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.FieldParamsItem"
                    }
                },
                "pagination": {
                    "description": "分页参数",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Pagination"
                        }
                    ]
                },
                "table_name": {
                    "type": "string"
                }
            }
        },
...
        "model.Ticket": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "details": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "description": "TicketProcessing TicketClosed",
                    "type": "string"
                },
                "ticket_message": {
                    "description": "一对多关联",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.TicketMessage"
                    }
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "model.TicketMessage": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_admin": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                },
                "ticket_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "response.ResponseStruct": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        }
    }
}